LPS-179567 Extract CPDefinitionPricingClassesRelationsScreenNavigationCategory methods to a CPDefinitionPricingClassesRelationsScreenNavigationEntry component and extends CPDefinitionPricingClassesRelationsScreenNavigationCategory in it
