+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/servlet/taglib/CPDefinitionPricingClassesRelationsScreenNavigationCategory.java
@@ -14,37 +14,13 @@
 
 package com.liferay.commerce.pricing.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.pricing.service.CommercePricingClassService;
-import com.liferay.commerce.pricing.web.internal.display.context.CPDefinitionPricingClassDisplayContext;
-import com.liferay.commerce.product.model.CPDefinition;
-import com.liferay.commerce.product.model.CommerceCatalog;
-import com.liferay.commerce.product.portlet.action.ActionHelper;
 import com.liferay.commerce.product.servlet.taglib.ui.constants.CPDefinitionScreenNavigationConstants;
-import com.liferay.commerce.product.type.CPType;
-import com.liferay.commerce.product.type.CPTypeRegistry;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -53,23 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Riccardo Alberti
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=85",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=85",
+	service = ScreenNavigationCategory.class
 )
 public class CPDefinitionPricingClassesRelationsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<CPDefinition> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CPDefinitionScreenNavigationConstants.
-			CATEGORY_KEY_PRICING_CLASSES;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CPDefinitionScreenNavigationConstants.
 			CATEGORY_KEY_PRICING_CLASSES;
 	}
@@ -79,7 +46,7 @@ public class CPDefinitionPricingClassesRelationsScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(
+		return language.get(
 			resourceBundle,
 			CPDefinitionScreenNavigationConstants.CATEGORY_KEY_PRICING_CLASSES);
 	}
@@ -90,83 +57,7 @@ public class CPDefinitionPricingClassesRelationsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_CP_DEFINITION_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CPDefinition cpDefinition) {
-		if (cpDefinition == null) {
-			return false;
-		}
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		CPType cpType = _cpTypeRegistry.getCPType(
-			cpDefinition.getProductTypeName());
-
-		try {
-			if (_commerceCatalogModelResourcePermission.contains(
-					permissionChecker, cpDefinition.getCommerceCatalog(),
-					ActionKeys.VIEW) &&
-				cpType.isProductGroupsEnabled()) {
-
-				return true;
-			}
-		}
-		catch (PortalException portalException) {
-			_log.error(portalException);
-		}
-
-		return false;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		CPDefinitionPricingClassDisplayContext
-			cpDefinitionPricingClassDisplayContext =
-				new CPDefinitionPricingClassDisplayContext(
-					_actionHelper, httpServletRequest,
-					_commerceCatalogModelResourcePermission,
-					_commercePricingClassService);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			cpDefinitionPricingClassDisplayContext);
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/cp_definitions/cp_definition_pricing_classes.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CPDefinitionPricingClassesRelationsScreenNavigationCategory.class);
-
 	@Reference
-	private ActionHelper _actionHelper;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.product.model.CommerceCatalog)"
-	)
-	private ModelResourcePermission<CommerceCatalog>
-		_commerceCatalogModelResourcePermission;
-
-	@Reference
-	private CommercePricingClassService _commercePricingClassService;
-
-	@Reference
-	private CPTypeRegistry _cpTypeRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.pricing.web)"
-	)
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
